Beta Final for Submission
Beta Final, The game is not fully working yet. Spent too much time trying to fix characters on Blender and on the Enemy AI's Ray tracing effects in the unreal blueprints, which I could not get working in the end.  I did manage to meet 9 of the twelve deliverables for this project, the missing three were using the sequencer, the menu screens and a saving and loading system.